requirements.txt
fastapi
requests

uvicorn

python-dotenv

aiohttp

openai

exceptiongroup

mcp[cli]==1.9.4


appbuilder-sdk

aiosqlite